# generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Working sample of Spring Boot for cars annotations",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Spring Boot for cars",
    "termsOfService" : "http://www.githubproeas.com",
    "contact" : {
      "name" : "Cars Example",
      "url" : "http",
      "email" : "carsexample@email.com"
    },
    "license" : {
      "name" : "License name",
      "url" : "http://www.license.com"
    }
  },
  "host" : "localhost:8081",
  "basePath" : "/",
  "schemes" : [ "http" ],
  "paths" : {
    "/cars" : {
      "post" : {
        "operationId" : "createCar",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Car"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Car"
            }
          }
        }
      }
    },
    "/cars/delete/{idCar}" : {
      "post" : {
        "operationId" : "deleteCarFromService",
        "parameters" : [ {
          "name" : "idCar",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    },
    "/cars/user/{id}" : {
      "get" : {
        "operationId" : "listByUserId",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Car" : {
      "type" : "object",
      "required" : [ "id_car", "id_user" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "id_car" : {
          "type" : "integer",
          "format" : "int64"
        },
        "id_user" : {
          "type" : "integer",
          "format" : "int64"
        },
        "brand" : {
          "type" : "string"
        },
        "model" : {
          "type" : "string"
        },
        "is_default" : {
          "type" : "boolean"
        },
        "created_at" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    }
  }
}